feat: Phase 6c — Cash Fixed entry polish + decimal keypad
Strict-design alignment for the Flow 4 cash-amount entry screen (vs
Screenshot 2026-05-11 225053.png), the last customer screen not yet on
the dispensing-family language.

- In-card StateChip + label header, and the ledger (Price / L · Cutoff ·
  Min–Max) now sits in the state-tinted rounded panel the other screens use.
- Keypad gains a decimal key (bottom-left, ✓ dropped — commit is the bottom
  AUTHORISE button), so the attendant can enter a sub-naira amount. Matches
  the design's own `parseFloat(cashAmount)` pseudo-code.
- Cash entry is now kobo-precise: typed naira (with optional kobo) is parsed
  and authorised as kobo; onCashFixedAuthorise takes Long kobo end to end.
- Extracted appendDigit/appendDecimal to ui/util/AmountInput so the pre-pay
  and cash keypads share one decimal-entry implementation.

+++ app/src/main/java/app/balancee/smartpump/display/ui/customer/CustomerStateHost.kt
@@ -46,7 +46,7 @@ fun CustomerStateHost(
     onAmountTileTap: (Int) -> Unit,
     onMethodTileTap: (PaymentMethod) -> Unit,
     onModeConfirm: () -> Unit,
-    onCashFixedAuthorise: (Int) -> Unit,
+    onCashFixedAuthorise: (Long) -> Unit,
     onFillupSelectIntent: (PostFillIntent) -> Unit,
     onFillupPayCash: () -> Unit,
     onFillupPayDigital: () -> Unit,
